requirements.txt
python-gnupg>=0.5.0
pyperclip>=1.8.2
cryptography>=41.0.0
windows-curses; platform_system=="Windows" 